btrfs-progs: move extent-cache.c to common/
