+++ check/common.h
@@ -22,7 +22,7 @@
 
 #include "kerncompat.h"
 #include "ctree.h"
-#include "extent-cache.h"
+#include "common/extent-cache.h"
 #include "kernel-lib/list.h"
 
 struct block_group_record {
