+++ cmds/rescue-chunk-recover.c
@@ -32,7 +32,7 @@
 #include "kernel-lib/list.h"
 #include "kernel-lib/radix-tree.h"
 #include "ctree.h"
-#include "extent-cache.h"
+#include "common/extent-cache.h"
 #include "disk-io.h"
 #include "volumes.h"
 #include "transaction.h"
