+++ common/extent-cache.c
@@ -18,7 +18,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include "kerncompat.h"
-#include "extent-cache.h"
+#include "common/extent-cache.h"
 #include "common/rbtree-utils.h"
 
 struct cache_extent_search_range {
