+++ tests/library-test.c
@@ -29,7 +29,7 @@
 #include "ioctl.h"
 #include "btrfs-list.h"
 #include "check/btrfsck.h"
-#include "extent-cache.h"
+#include "common/extent-cache.h"
 #include "send.h"
 #include "send-stream.h"
 #include "send-utils.h"
